refactor: extract the command line into src/cli.ts
index.ts runs main() at import time unless SSH_MCP_DISABLE_MAIN=1, so every test
of a pure argv helper was a test that could start connecting to real hosts on a
machine with a real config. config-failure-reporting.test.ts carried a vi.stubEnv
and a six-line comment about exactly that; without it, a single-file run reported
passing tests alongside "process.exit unexpectedly called with 1".

parseArgv, flagEnabled, REMOVED_V1_FLAGS/checkRemovedFlags, parseMaxChars,
resolveHostKeyMode, resolveHostGroup and buildAppConfig — with the aclFindings
array buildAppConfig fills — move to src/cli.ts, which has no import-time side
effect. index.ts keeps its imports, main() and the boot gate, and drops from 369
lines to 154, under the 200 the repo asks of an entry point.

No back-compat re-export. The trailing `export { … }` block existed only for
tests, and leaving it would have preserved the hazard the split removes: the four
files that imported it now import cli.js, and the scaffolding that existed solely
for the boot gate is gone with it. The resetModules dance in
config-failure-reporting stays — that one is about `instanceof` across a module
reset, a different reason.

Pure move, verified as one: the relocated block is byte-for-byte identical to the
original apart from six `export` prefixes (checked line by line against
HEAD:src/index.ts), no test body changed, and `--version`/`--help` behave exactly
as they do on main — both already print the config error there, which is the known
gap queued under #41 rather than anything this touched.

No changeset: the package publishes only build/ behind a bin, with no main,
exports or types, so no consumer can observe the file layout.

Verified: 659 tests against the live containers, 29 e2e.
